requirements.txt
beautifulsoup4==4.14.3
crewai==1.13.0
langchain_google_genai==4.2.1
langchain_tavily==0.2.17
langgraph==1.1.6
litellm==1.83.3
numpy==2.4.4
nvdlib==0.8.3
pandas==3.0.2
protobuf==7.34.1
pydantic==2.12.5
python-dotenv==1.2.2
python_nmap==0.7.1
python_pptx==1.0.2
Requests==2.33.1
shodan==1.31.0
streamlit==1.47.1
tavily_python==0.7.11
fastapi
uvicorn[standard]
openpyxl
redis
python-multipart